// src/generated/t_tech/invest/grpc/marketdata.ts
export enum CandleInterval {
  /** CANDLE_INTERVAL_UNSPECIFIED - Интервал не определен. */
  CANDLE_INTERVAL_UNSPECIFIED = 0,
  /** CANDLE_INTERVAL_1_MIN - От 1 минуты до 1 дня. Максимальное значение `limit` — 2400. */
  CANDLE_INTERVAL_1_MIN = 1,
  /** CANDLE_INTERVAL_5_MIN - От 5 минут до недели. Максимальное значение `limit` — 2400. */
  CANDLE_INTERVAL_5_MIN = 2,
  /** CANDLE_INTERVAL_15_MIN - От 15 минут до 3 недель. Максимальное значение `limit` — 2400. */
  CANDLE_INTERVAL_15_MIN = 3,
  /** CANDLE_INTERVAL_HOUR - От 1 часа до 3 месяцев. Максимальное значение `limit` — 2400. */
  CANDLE_INTERVAL_HOUR = 4,
  /** CANDLE_INTERVAL_DAY - От 1 дня до 6 лет. Максимальное значение `limit` — 2400. */
  CANDLE_INTERVAL_DAY = 5,
  /** CANDLE_INTERVAL_2_MIN - От 2 минут до 1 дня. Максимальное значение `limit` — 1200. */
  CANDLE_INTERVAL_2_MIN = 6,
  /** CANDLE_INTERVAL_3_MIN - От 3 минут до 1 дня. Максимальное значение `limit` — 750. */
  CANDLE_INTERVAL_3_MIN = 7,
  /** CANDLE_INTERVAL_10_MIN - От 10 минут до недели. Максимальное значение `limit` — 1200. */
  CANDLE_INTERVAL_10_MIN = 8,
  /** CANDLE_INTERVAL_30_MIN - От 30 минут до 3 недель. Максимальное значение `limit` — 1200. */
  CANDLE_INTERVAL_30_MIN = 9,
  /** CANDLE_INTERVAL_2_HOUR - От 2 часов до 3 месяцев. Максимальное значение `limit` — 2400. */
  CANDLE_INTERVAL_2_HOUR = 10,
  /** CANDLE_INTERVAL_4_HOUR - От 4 часов до 3 месяцев. Максимальное значение `limit` — 700. */
  CANDLE_INTERVAL_4_HOUR = 11,
  /** CANDLE_INTERVAL_WEEK - От 1 недели до 5 лет. Максимальное значение `limit` — 300. */
  CANDLE_INTERVAL_WEEK = 12,
  /** CANDLE_INTERVAL_MONTH - От 1 месяца до 10 лет. Максимальное значение `limit` — 120. */
  CANDLE_INTERVAL_MONTH = 13,
  /** CANDLE_INTERVAL_5_SEC - От 5 секунд до 200 минут. Максимальное значение `limit` — 2500. */
  CANDLE_INTERVAL_5_SEC = 14,
  /** CANDLE_INTERVAL_10_SEC - От 10 секунд до 200 минут. Максимальное значение `limit` — 1250. */
  CANDLE_INTERVAL_10_SEC = 15,
  /** CANDLE_INTERVAL_30_SEC - От 30 секунд до 20 часов. Максимальное значение `limit` — 2500. */
  CANDLE_INTERVAL_30_SEC = 16,
  UNRECOGNIZED = -1,
}
// ...
export interface GetCandlesRequest {
  /**
   * Deprecated FIGI-идентификатор инструмента. Используйте `instrument_id`.
   *
   * @deprecated
   */
  figi?:
    | string
    | undefined;
  /** Начало запрашиваемого периода по UTC. */
  from:
    | Date
    | undefined;
  /** Окончание запрашиваемого периода по UTC. */
  to:
    | Date
    | undefined;
  /** Интервал запрошенных свечей. */
  interval: CandleInterval;
  /** Идентификатор инструмента. Принимает значение `figi`, `instrument_uid` или `ticker + '_' + class_code`. */
  instrumentId?:
    | string
    | undefined;
  /** Тип источника свечи. */
  candleSourceType?:
    | GetCandlesRequest_CandleSource
    | undefined;
  /** Максимальное количество свечей в ответе. */
  limit?: number | undefined;
}

export enum GetCandlesRequest_CandleSource {
  /** CANDLE_SOURCE_UNSPECIFIED - Все свечи. */
  CANDLE_SOURCE_UNSPECIFIED = 0,
  /** CANDLE_SOURCE_EXCHANGE - Биржевые свечи. */
  CANDLE_SOURCE_EXCHANGE = 1,
  /** CANDLE_SOURCE_INCLUDE_WEEKEND - Все свечи с учетом торговли по выходным. */
  CANDLE_SOURCE_INCLUDE_WEEKEND = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface GetCandlesResponse {
  /** Массив свечей. */
  candles: HistoricCandle[];
}

/** Информация о свече. */
export interface HistoricCandle {
  /** Цена открытия за 1 инструмент. Чтобы получить стоимость лота, нужно умножить на лотность инструмента. [Подробнее про перевод цен в валюту](./faq_marketdata/#_15). */
  open:
    | Quotation
    | undefined;
  /** Максимальная цена за 1 инструмент. Чтобы получить стоимость лота, нужно умножить на лотность инструмента. [Подробнее про перевод цен в валюту](./faq_marketdata/#_15). */
  high:
    | Quotation
    | undefined;
  /** Минимальная цена за 1 инструмент. Чтобы получить стоимость лота, нужно умножить на лотность инструмента. [Подробнее про перевод цен в валюту](./faq_marketdata/#_15). */
  low:
    | Quotation
    | undefined;
  /** Цена закрытия за 1 инструмент. Чтобы получить стоимость лота, нужно умножить на лотность инструмента. [Подробнее про перевод цен в валюту](./faq_marketdata/#_15). */
  close:
    | Quotation
    | undefined;
  /** Объем торгов в лотах. */
  volume: number;
  /** Время свечи в часовом поясе UTC. */
  time:
    | Date
    | undefined;
  /** Признак завершенности свечи. **false** — свеча за текущие интервал еще сформирована не полностью. */
  isComplete: boolean;
  /** Тип источника свечи */
  candleSource: CandleSource;
  /** Объем торгов на покупку. */
  volumeBuy: number;
  /** Объём торгов на продажу. */
  volumeSell: number;
}
// ...
export const GetCandlesRequest: MessageFns<GetCandlesRequest> = {
  encode(message: GetCandlesRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.figi !== undefined) {
      writer.uint32(10).string(message.figi);
    }
    if (message.from !== undefined) {
      Timestamp.encode(toTimestamp(message.from), writer.uint32(18).fork()).join();
    }
    if (message.to !== undefined) {
      Timestamp.encode(toTimestamp(message.to), writer.uint32(26).fork()).join();
    }
    if (message.interval !== 0) {
      writer.uint32(32).int32(message.interval);
    }
    if (message.instrumentId !== undefined) {
      writer.uint32(42).string(message.instrumentId);
    }
    if (message.candleSourceType !== undefined) {
      writer.uint32(56).int32(message.candleSourceType);
    }
    if (message.limit !== undefined) {
      writer.uint32(80).int32(message.limit);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetCandlesRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetCandlesRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.figi = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.from = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.to = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.interval = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.instrumentId = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 56) {
            break;
          }

          message.candleSourceType = reader.int32() as any;
          continue;
        }
        case 10: {
          if (tag !== 80) {
            break;
          }

          message.limit = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetCandlesRequest {
    return {
      figi: isSet(object.figi) ? globalThis.String(object.figi) : undefined,
      from: isSet(object.from) ? fromJsonTimestamp(object.from) : undefined,
      to: isSet(object.to) ? fromJsonTimestamp(object.to) : undefined,
      interval: isSet(object.interval) ? candleIntervalFromJSON(object.interval) : 0,
      instrumentId: isSet(object.instrumentId)
        ? globalThis.String(object.instrumentId)
        : isSet(object.instrument_id)
        ? globalThis.String(object.instrument_id)
        : undefined,
      candleSourceType: isSet(object.candleSourceType)
        ? getCandlesRequest_CandleSourceFromJSON(object.candleSourceType)
        : isSet(object.candle_source_type)
        ? getCandlesRequest_CandleSourceFromJSON(object.candle_source_type)
        : undefined,
      limit: isSet(object.limit) ? globalThis.Number(object.limit) : undefined,
    };
  },

  toJSON(message: GetCandlesRequest): unknown {
    const obj: any = {};
    if (message.figi !== undefined) {
      obj.figi = message.figi;
    }
    if (message.from !== undefined) {
      obj.from = message.from.toISOString();
    }
    if (message.to !== undefined) {
      obj.to = message.to.toISOString();
    }
    if (message.interval !== 0) {
      obj.interval = candleIntervalToJSON(message.interval);
    }
    if (message.instrumentId !== undefined) {
      obj.instrumentId = message.instrumentId;
    }
    if (message.candleSourceType !== undefined) {
      obj.candleSourceType = getCandlesRequest_CandleSourceToJSON(message.candleSourceType);
    }
    if (message.limit !== undefined) {
      obj.limit = Math.round(message.limit);
    }
    return obj;
  },

  create(base?: DeepPartial<GetCandlesRequest>): GetCandlesRequest {
    return GetCandlesRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetCandlesRequest>): GetCandlesRequest {
    const message = createBaseGetCandlesRequest();
    message.figi = object.figi ?? undefined;
    message.from = object.from ?? undefined;
    message.to = object.to ?? undefined;
    message.interval = object.interval ?? 0;
    message.instrumentId = object.instrumentId ?? undefined;
    message.candleSourceType = object.candleSourceType ?? undefined;
    message.limit = object.limit ?? undefined;
    return message;
  },
};

function createBaseGetCandlesResponse(): GetCandlesResponse {
  return { candles: [] };
}

export const GetCandlesResponse: MessageFns<GetCandlesResponse> = {
  encode(message: GetCandlesResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.candles) {
      HistoricCandle.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetCandlesResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetCandlesResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.candles.push(HistoricCandle.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetCandlesResponse {
    return {
      candles: globalThis.Array.isArray(object?.candles)
        ? object.candles.map((e: any) => HistoricCandle.fromJSON(e))
        : [],
    };
  },

  toJSON(message: GetCandlesResponse): unknown {
    const obj: any = {};
    if (message.candles?.length) {
      obj.candles = message.candles.map((e) => HistoricCandle.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<GetCandlesResponse>): GetCandlesResponse {
    return GetCandlesResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetCandlesResponse>): GetCandlesResponse {
    const message = createBaseGetCandlesResponse();
    message.candles = object.candles?.map((e) => HistoricCandle.fromPartial(e)) || [];
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}